requirements.txt
streamlit
crewai
chromadb
langchain
openai
tiktoken
python-dotenv
sentence-transformers
selenium
webdriver-manager
google-generativeai